requirements.txt
streamlit
ollama
fastapi
uvicorn
python-dotenv
openai
jinja2